Fix an issue in executing regutil
`regutil` expects an option string, which depends on the types of input
data.  For example, the option string is `-s` for a string and `-u` for
an unsigned integer.  This commit split the task executing `regutil`
into two: one expects that the type on input data is a string, and the
other expects that the type of input data is an unsigned integer.